requirements.txt
asgiref==3.8.1
click==8.1.7
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
h11==0.14.0
packaging==24.1
pillow==11.0.0
psycopg==3.2.3
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
uvicorn==0.32.0
